Use authprovider/update context for connection type
